Extract pandaops middleware execution methods to `ExecutionApi` struct.
