+++ portal-bridge/src/lib.rs
@@ -6,6 +6,7 @@ pub mod cli;
 pub mod client_handles;
 pub mod consensus_api;
 pub mod constants;
+pub mod execution_api;
 pub mod full_header;
 pub mod mode;
 pub mod pandaops;
